Extract reactor
