+++ packages/vfx-composer/src/fiber/index.ts
@@ -2,3 +2,4 @@ export * from "./Emitter"
 export * from "./makeParticles"
 export * from "./Particles"
 export * from "./VFXMaterial"
+export { VFXReactor as VFX } from "./reactor"
